+++ src/it/osgi/src/main/java/org/bytedeco/javacpp/test/osgi/JavaCPPOsgiTest.java
@@ -9,5 +9,8 @@ public class JavaCPPOsgiTest {
     @Test
     public void testJavaCPP() {
         assertEquals(3, Calc.add(1, 2));
+
+        Calc.IntValue v = new Calc.IntValue().value(4);
+        assertEquals(4, v.value());
     }
 }