+++ src/main/java/org/bytedeco/javacpp/package-info.java
@@ -2,5 +2,4 @@
  * Contains the main set of classes for JavaCPP at runtime.
  */
 @org.osgi.annotation.bundle.Export
-@org.osgi.annotation.versioning.Version("1.5.0")
 package org.bytedeco.javacpp;