+++ src/main/java/org/bytedeco/javacpp/tools/package-info.java
@@ -3,5 +3,4 @@
  */
 @org.osgi.annotation.bundle.Export(
 		attribute="exclude:=\"BuildMojo,ParseMojo,CacheMojo\"")
-@org.osgi.annotation.versioning.Version("1.5.0")
 package org.bytedeco.javacpp.tools;